fix: derive authoritative catalog drain budget from protocol lanes

+++ src/Win7POS.Core/Online/OnlineSyncSupervisorContracts.cs
@@ -142,7 +142,7 @@ namespace Win7POS.Core.Online
             bool requestCatalogNow = false,
             int? nextPollAfterSeconds = null,
             bool terminal = false,
-            int catalogPagesProcessed = 0,
+            long catalogPagesProcessed = 0,
             int catalogRowsApplied = 0,
             bool catalogSaleSafe = false,
             PosRuntimeDiagnostic catalogDiagnostic = null)
@@ -173,7 +173,7 @@ namespace Win7POS.Core.Online
 
         public bool AuthenticationDenied { get; }
         public bool CatalogHasMore { get; }
-        public int CatalogPagesProcessed { get; }
+        public long CatalogPagesProcessed { get; }
         public int CatalogRowsApplied { get; }
         public bool CatalogSaleSafe { get; }
         public PosRuntimeDiagnostic CatalogDiagnostic { get; }
