+++ src/Win7POS.Data/Online/CatalogSyncCoordinator.cs
@@ -49,7 +49,7 @@ namespace Win7POS.Data.Online
             bool offline = false,
             bool hasMore = false,
             bool receivedChanges = false,
-            int pages = 0,
+            long pages = 0,
             int rows = 0,
             long durationMilliseconds = 0,
             string resumeCursor = null,
@@ -92,7 +92,7 @@ namespace Win7POS.Data.Online
         public long? NextOutboxRetryAt { get; }
         public bool Offline { get; }
         public bool OutboxWorkRemaining { get; }
-        public int Pages { get; }
+        public long Pages { get; }
         public bool ReceivedChanges { get; }
         public string ResumeCursor { get; }
         public int Rows { get; }
